Clothing initiative to help underprivelged individuals feel confident to attend job interviews
Our Grassroots Clothing Bank is in it's infancy, but we are hoping it will grow and GROW!!!
The idea is to give people an inner confidence when offered job interviews by hopefully being able to give a good first impression, especially when many may not have attended a job interview for many years.

This idea has developed due to a lack of this type of service in the community and we welcome donations from individuals and organisations alike, who feel that they would like to donate suitable "interview styled clothing".  If interested in helping us, please call our East Hub Office in Craigmillar, Edinburgh on 0131 629 9858.